Create Simon tests
Messing with how to create a circle and then place things correctly within that circle. 
Can create circle, but children divs are not being contained and flowing outside the container. 
Will continue to work on positioning within circle while simultaneously trying to work out how to create the shape of each colored button in Simon (The kind of crescent shape).