Updated HTML/CSS
Added new div that is the inner circle, which will have the buttons and counter on it. InnerCircle div was used to give the correct shape to color divs. 
Changed CSS to give correct appearance in all resolutions and changed each color div correct color and shape. Will change to buttons and add innerCircle divs next