Updated JS
playerArray is now reset at the end of computerTurn in order to let player hit all the same buttons. 
Ultimately, would be MUCH more efficient to have final array checked to the computer Array, but will continue from here and move on (for loop should be replaced with if (playerArray == computerArray), but need to figure out how to make sure those are equivalent.